# backend-rails/openapi/v1.yaml
openapi: 3.0.3
info:
  title: School Election API
  version: 1.0.0
  description: >-
    Versioned JSON contract for the separate frontend. Operations marked planned
    are design targets, not available routes. Their payloads remain draft until
    the corresponding feature is implemented and covered by request tests.
servers:
  - url: /
paths:
...
  /api/v1/admin/rounds/{id}/open:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    post:
      operationId: openRound
      summary: Freeze a valid round configuration and open consecutive voting stages
      x-implementation-status: implemented
      x-authentication: creator-session-with-csrf
      responses:
        '200':
          description: Opened round state, snapshot digest and ordered stages
          content:
            application/json:
              schema:
                type: object
        default:
          $ref: '#/components/responses/ApiError'
  /api/v1/admin/rounds/{id}/suspend:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    post:
      operationId: suspendRound
      summary: Suspend voting while preserving confirmed votes and session progress
      description: >-
        Requires an active creator of this election and a nonblank textual reason.
        New releases and confirmations are blocked. Previously committed confirmations
        remain recoverable. State, incident and operational audit are persisted atomically.
      x-implementation-status: implemented
      x-authentication: creator-session-with-csrf
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoundTransitionRequest'
      responses:
        '200':
          description: Suspended round
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoundTransitionResult'
        '401':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
        '409':
          description: round_suspend_denied; the round is not open
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: invalid_reason; a nonblank textual reason is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/admin/rounds/{id}/resume:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    post:
      operationId: resumeRound
      summary: Resume the same ballot within the original voting schedule
      description: >-
        Requires an active creator of this election and a nonblank textual reason.
        Only a suspended round can resume, from opens_at inclusive until grace_until
        exclusive. The original schedule is not extended. After closes_at, only
        sessions already started can continue; new releases and starts remain blocked.
        State, incident and operational audit are persisted atomically.
      x-implementation-status: implemented
      x-authentication: creator-session-with-csrf
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoundTransitionRequest'
      responses:
        '200':
          description: Resumed round
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoundTransitionResult'
        '401':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
        '409':
          description: round_resume_denied; incompatible state or original grace period ended
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: invalid_reason; a nonblank textual reason is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/admin/rounds/{id}/close:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    post:
      operationId: closeRound
      summary: Close after the grace period and persist reconciled tallies
      x-implementation-status: implemented
      x-authentication: creator-session-with-csrf
      responses:
        '200':
          description: Closed round and final or pending tally states
          content:
            application/json:
              schema:
                type: object
        default:
          $ref: '#/components/responses/ApiError'
...
components:
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    ApiError:
      description: Stable JSON error; the HTTP status identifies the failure class
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: not_found
            message:
              type: string
              example: Resource not found
            details:
              type: array
              items:
                type: object
...
    RoundTransitionRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 1
          pattern: '\S'
          description: Operational justification; never include a voter identity or choice.
    RoundTransitionResult:
      type: object
      required: [state]
      properties:
        state:
          type: string
          enum: [open, suspended]
...